# Insight-Performance-Hub/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Organisational Performance Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scorecards:
    get:
      operationId: listScorecards
      tags: [scorecards]
      summary: List scorecards
      parameters:
        - name: cycleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Scorecard list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scorecard"
    post:
      operationId: createScorecard
      tags: [scorecards]
      summary: Create a scorecard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScorecardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"
  /scorecards/{id}:
    get:
      operationId: getScorecard
      tags: [scorecards]
      summary: Get a scorecard by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scorecard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"
    patch:
      operationId: updateScorecard
      tags: [scorecards]
      summary: Update a scorecard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScorecardInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"
...
  /dept-scorecards:
    get:
      operationId: listDeptScorecards
      tags: [deptScorecards]
      summary: List departmental scorecards
      parameters:
        - name: cycleId
          in: query
          schema:
            type: integer
        - name: departmentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of departmental scorecards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeptScorecard"
    post:
      operationId: createDeptScorecard
      tags: [deptScorecards]
      summary: Create departmental scorecard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeptScorecardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeptScorecard"
  /dept-scorecards/{id}:
    patch:
      operationId: updateDeptScorecard
      tags: [deptScorecards]
      summary: Update departmental scorecard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeptScorecardInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeptScorecard"
  /dept-scorecards/{id}/transition:
    post:
      operationId: transitionDeptScorecard
      tags: [deptScorecards]
      summary: Transition departmental scorecard workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeptScorecardTransitionInput"
      responses:
        "200":
          description: Transitioned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeptScorecard"
  /dept-scorecards/{id}/inherit-kpis:
    post:
      operationId: inheritKpisFromOrg
      tags: [deptScorecards]
      summary: Inherit KPIs from organisational scorecard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inherited KPIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeptScorecardKpi"
  /dept-scorecards/{deptScorecardId}/kpis:
    get:
      operationId: listDeptScorecardKpis
      tags: [deptScorecards]
      summary: List KPIs for departmental scorecard
      parameters:
        - name: deptScorecardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of KPIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeptScorecardKpi"
    post:
      operationId: createDeptScorecardKpi
      tags: [deptScorecards]
      summary: Add KPI to departmental scorecard
      parameters:
        - name: deptScorecardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDeptKpiInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeptScorecardKpi"
  /dept-kpis/{id}:
    patch:
      operationId: updateDeptKpi
      tags: [deptScorecards]
      summary: Update departmental KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeptKpiInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeptScorecardKpi"
    delete:
      operationId: deleteDeptKpi
      tags: [deptScorecards]
      summary: Delete departmental KPI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Scorecard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cycleId:
          type: integer
        scorecardType:
          type: string
        departmentId:
          type: integer
          nullable: true
        status:
          type: string
        approvedById:
          type: integer
          nullable: true
        approvedAt:
          type: string
          nullable: true
        approvalComments:
          type: string
          nullable: true
        returnComments:
          type: string
          nullable: true
        createdById:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - cycleId
        - scorecardType
        - status
        - createdById
    CreateScorecardInput:
      type: object
      properties:
        name:
          type: string
        cycleId:
          type: integer
        scorecardType:
          type: string
          default: organisational
        departmentId:
          type: integer
      required:
        - name
        - cycleId
    UpdateScorecardInput:
      type: object
      properties:
        name:
          type: string
        scorecardType:
          type: string
        departmentId:
          type: integer
...
    DeptScorecard:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cycleId:
          type: integer
        departmentId:
          type: integer
        departmentName:
          type: string
        parentScorecardId:
          type: integer
          nullable: true
        ownerId:
          type: integer
          nullable: true
        status:
          type: string
        approvedById:
          type: integer
          nullable: true
        approvedAt:
          type: string
          nullable: true
        approvalComments:
          type: string
          nullable: true
        createdById:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - cycleId
        - departmentId
        - departmentName
        - status
        - createdById
    CreateDeptScorecardInput:
      type: object
      properties:
        name:
          type: string
        cycleId:
          type: integer
        departmentId:
          type: integer
        departmentName:
          type: string
        parentScorecardId:
          type: integer
        ownerId:
          type: integer
      required:
        - name
        - cycleId
        - departmentId
        - departmentName
    UpdateDeptScorecardInput:
      type: object
      properties:
        name:
          type: string
        departmentName:
          type: string
        ownerId:
          type: integer
    DeptScorecardTransitionInput:
      type: object
      properties:
        action:
          type: string
          enum: [submit, approve, return, lock]
        comments:
          type: string
      required:
        - action
    DeptScorecardKpi:
      type: object
      properties:
        id:
          type: integer
        deptScorecardId:
          type: integer
        parentKpiId:
          type: integer
          nullable: true
        kpiNumber:
          type: string
        description:
          type: string
        strategicObjective:
          type: string
          nullable: true
        nkpaLink:
          type: string
          nullable: true
        responsiblePostId:
          type: integer
          nullable: true
        baseline:
          type: string
          nullable: true
        annualTarget:
          type: string
        annualBudgetTarget:
          type: number
          nullable: true
        weighting:
          type: number
        unitOfMeasureId:
          type: integer
          nullable: true
        isCumulative:
          type: boolean
        customFields:
          type: object
          additionalProperties: true
          nullable: true
        isInherited:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - deptScorecardId
        - kpiNumber
        - description
        - annualTarget
        - weighting
    CreateDeptKpiInput:
      type: object
      properties:
        parentKpiId:
          type: integer
        kpiNumber:
          type: string
        description:
          type: string
        strategicObjective:
          type: string
        nkpaLink:
          type: string
        responsiblePostId:
          type: integer
        baseline:
          type: string
        annualTarget:
          type: string
        annualBudgetTarget:
          type: number
        weighting:
          type: number
        unitOfMeasureId:
          type: integer
        isCumulative:
          type: boolean
        customFields:
          type: object
          additionalProperties: true
          nullable: true
      required:
        - kpiNumber
        - description
        - annualTarget
    UpdateDeptKpiInput:
      type: object
      properties:
        kpiNumber:
          type: string
        description:
          type: string
        strategicObjective:
          type: string
        nkpaLink:
          type: string
        responsiblePostId:
          type: integer
        baseline:
          type: string
        annualTarget:
          type: string
        annualBudgetTarget:
          type: number
        weighting:
          type: number
        unitOfMeasureId:
          type: integer
        isCumulative:
          type: boolean
        customFields:
          type: object
          additionalProperties: true
          nullable: true
...